fix bugs and change routes

+++ app/controllers/friend_controller.rb
@@ -28,7 +28,7 @@ class FriendController < ApplicationController
 		f.update(status: true)
 		@friend = params[:friend_id]
 
-		redirect_to friendprofile_path(key: params[:friend_id]), notice: "Freund wurde hinzugefügt"
+		redirect_to friendprofile_path(params[:friend_id]), notice: "Freund wurde hinzugefügt"
 	end
 
 	def frienddenie
@@ -42,11 +42,12 @@ class FriendController < ApplicationController
 	def frienddelete
 
 		c = Conversation.where("(sender_id = ? AND reciepient_id = ?) OR (sender_id = ? AND reciepient_id = ?)",current_user.id,  params[:friend_id], params[:friend_id], current_user.id ).first
-		m = Message.where(conversation_id: c.id)	
+		unless c.nil?
+			m = Message.where(conversation_id: c.id)
+			m.destroy_all
+			c.destroy	
+		end
 		f = Friendship.where("user_id = ? AND friend_id = ? OR friend_id = ? AND user_id = ?",current_user , params[:friend_id],current_user, params[:friend_id]).first 
-		
-		m.destroy_all
-		c.destroy
 	    f.destroy
 
 		redirect_to friend_path, notice: "Freund wurde erfolgreich gelöscht"
@@ -62,7 +63,7 @@ class FriendController < ApplicationController
 
 		Pinboard.create(user_id: current_user.id, friend_id: params[:friend_id], entry: params[:msg])
 
-		redirect_to friendprofile_path(key: params[:friend_id]), notice: "Pinwandeintrag wurde erfolgreich erstellt"
+		redirect_to friendprofile_path(params[:friend_id]), notice: "Pinwandeintrag wurde erfolgreich erstellt"
 	end
 
 	def frienddeletepinentry
